refactor: meal section에서 기록 뷰 분리
